🎨: WASM PoC
This is a very simple PoC of running a wasm binary selectable from a N8N node.

The interface would need to be defined, this would probably by something like:
environment like envvars
args like command line args
stdin a data stream that should be written to the stdin of the wasm binary

To run it, link it, add a Node fill the WasmBinary in with `demo.wasm`, this is
our example wasm.

Change-type: patch
Signed-off-by: Andreas Fitzek <andreas.fitzek@gmail.com>